:recycle: refactor: Alter names from usePosts to usePostData
Besides the name, other variables names on the function were
also exchanged so that this custom hook could be more broadly used
without causing dissonance upon use.
This was done because there is no good reason to replicate
the code 1:1 just changing the name of the function given
that with a small change it serves both.